mid_pmu: otc ics pmu driver patches ported to umg ics
BZ: 34256

	commit c1e44f12ce693e1e49d409808a1da2de37182142
	Author: Olev Kartau <olev.kartau@intel.com>
	Date:   Thu Mar 15 15:04:03 2012 +0200

	    pmu: interrupt status read value should be signed int

	    u8 status was used to store value from pmu_read_interrupt_status
	    which returns signed int. Comparing u8 to -1 could never match.
	    Also, split status and ret.value to differently typed variables.


	commit b123e410c7ae3019b2eecc7c7549002be559d45a
	Author: Austin Zhang <austin.zhang@intel.com>
	Date:   Thu Feb 23 09:23:53 2012 -0800

	    KER-106: avoid s0i1-idle hang

	    Prevent re-entrant S0i1, just like we do for S0i3;
	    and thus avoid a system hang with S0i1 enabled.

	    https://sa.otcshare.org/jira/browse/KER-106

	    With this patch in place, and loglevel 15
	    (which works around GRA-27), S0i1 has never
	    been observed to hang during over-night testing.

	    Tested-by: Yong Wang <yong.y.wang@intel.com
	    Tested-by: Kristen C. Accardi <kristen.c.accardi@intel.com>

Reviewed-on: http://android.intel.com:8080/48298
Tested-by: Martin, LoicX <loicx.martin@intel.com>
Tested-by: buildbot <buildbot@intel.com>

+++ arch/x86/platform/intel-mid/intel_soc_pm_debug.c
@@ -54,7 +54,7 @@ static void pmu_log_timestamp(struct timespec *ts)
 	}
 }
 
-void pmu_log_pmu_irq(u8 status, bool interactive_cmd_sent)
+void pmu_log_pmu_irq(int status, bool interactive_cmd_sent)
 {
 	struct mid_pmu_pmu_irq_log *log =
 		&mid_pmu_cxt->pmu_irq_log[mid_pmu_cxt->pmu_irq_log_idx];
@@ -209,7 +209,7 @@ void pmu_dump_logs(void)
 	pmu_dump_ipc_irq_log();
 }
 #else
-void pmu_log_pmu_irq(u8 status, bool interactive_cmd_sent) {}
+void pmu_log_pmu_irq(int status, bool interactive_cmd_sent) {}
 void pmu_log_command(u32 command, struct pmu_ss_states *pm_ssc) {}
 void pmu_dump_logs(void) {}
 #endif /* LOG_PMU_EVENTS */
